refactor(web-scraping): rename interface files to responsibility-based names
scraping_interface.py → scraper.py
artifact_interface.py → artifact_repository.py

インターフェース命名規則の整理。クラス名 IScraper /
IArtifactRepository は変更なし。
