+++ lambda/web-scraping/src/domain/artifact_repository.py
