+++ lambda/web-scraping/src/domain/scraper.py
